[OJS.de lucene - part I - ap3] implemented interfaces ArticleSearchIndex<->LucenePlugin<->SolrWebService

+++ classes/search/ArticleSearchIndex.inc.php
@@ -31,7 +31,7 @@ class ArticleSearchIndex {
 	 */
 	function _indexObjectKeywords($objectId, $text, &$position) {
 		$searchDao =& DAORegistry::getDAO('ArticleSearchDAO');
-		$keywords =& ArticleSearchIndex::filterKeywords($text);
+		$keywords =& $this->filterKeywords($text);
 		for ($i = 0, $count = count($keywords); $i < $count; $i++) {
 			if ($searchDao->insertObjectKeyword($objectId, $keywords[$i], $position) !== null) {
 				$position += 1;
@@ -50,7 +50,7 @@ class ArticleSearchIndex {
 		$searchDao =& DAORegistry::getDAO('ArticleSearchDAO');
 		$objectId = $searchDao->insertObject($articleId, $type, $assocId);
 		$position = 0;
-		ArticleSearchIndex::_indexObjectKeywords($objectId, $text, $position);
+		$this->_indexObjectKeywords($objectId, $text, $position);
 	}
 
 	/**
@@ -60,24 +60,34 @@ class ArticleSearchIndex {
 	 * @param $fileId int
 	 */
 	function updateFileIndex($articleId, $type, $fileId) {
-		import('classes.file.ArticleFileManager');
-		$fileManager = new ArticleFileManager($articleId);
-		$file =& $fileManager->getFile($fileId);
+		// Check whether a search plug-in jumps in.
+		$hookResult =& HookRegistry::call(
+			'ArticleSearchIndex::updateFileIndex',
+			array($articleId, $type, $fileId)
+		);
 
-		if (isset($file)) {
-			$parser =& SearchFileParser::fromFile($file);
-		}
+		// If no search plug-in is activated then fall back to the
+		// default database search implementation.
+		if ($hookResult === false || is_null($hookResult)) {
+			import('classes.file.ArticleFileManager');
+			$fileManager = new ArticleFileManager($articleId);
+			$file =& $fileManager->getFile($fileId);
 
-		if (isset($parser)) {
-			if ($parser->open()) {
-				$searchDao =& DAORegistry::getDAO('ArticleSearchDAO');
-				$objectId = $searchDao->insertObject($articleId, $type, $fileId);
+			if (isset($file)) {
+				$parser =& SearchFileParser::fromFile($file);
+			}
 
-				$position = 0;
-				while(($text = $parser->read()) !== false) {
-					ArticleSearchIndex::_indexObjectKeywords($objectId, $text, $position);
+			if (isset($parser)) {
+				if ($parser->open()) {
+					$searchDao =& DAORegistry::getDAO('ArticleSearchDAO');
+					$objectId = $searchDao->insertObject($articleId, $type, $fileId);
+
+					$position = 0;
+					while(($text = $parser->read()) !== false) {
+						$this->_indexObjectKeywords($objectId, $text, $position);
+					}
+					$parser->close();
 				}
-				$parser->close();
 			}
 		}
 	}
@@ -89,8 +99,18 @@ class ArticleSearchIndex {
 	 * @param $assocId int optional
 	 */
 	function deleteTextIndex($articleId, $type = null, $assocId = null) {
-		$searchDao =& DAORegistry::getDAO('ArticleSearchDAO'); /* @var $searchDao ArticleSearchDAO */
-		return $searchDao->deleteArticleKeywords($articleId, $type, $assocId);
+		// Check whether a search plug-in jumps in.
+		$hookResult =& HookRegistry::call(
+			'ArticleSearchIndex::deleteTextIndex',
+			array($articleId, $type, $assocId)
+		);
+
+		// If no search plug-in is activated then fall back to the
+		// default database search implementation.
+		if ($hookResult === false || is_null($hookResult)) {
+			$searchDao =& DAORegistry::getDAO('ArticleSearchDAO'); /* @var $searchDao ArticleSearchDAO */
+			return $searchDao->deleteArticleKeywords($articleId, $type, $assocId);
+		}
 	}
 
 	/**
@@ -101,7 +121,7 @@ class ArticleSearchIndex {
 	 */
 	function &filterKeywords($text, $allowWildcards = false) {
 		$minLength = Config::getVar('search', 'min_word_length');
-		$stopwords =& ArticleSearchIndex::_loadStopwords();
+		$stopwords =& $this->_loadStopwords();
 
 		// Join multiple lines into a single string
 		if (is_array($text)) $text = join("\n", $text);
@@ -151,35 +171,45 @@ class ArticleSearchIndex {
 	 * @param $article Article
 	 */
 	function indexArticleMetadata(&$article) {
-		// Build author keywords
-		$authorText = array();
-		$authors = $article->getAuthors();
-		for ($i=0, $count=count($authors); $i < $count; $i++) {
-			$author =& $authors[$i];
-			array_push($authorText, $author->getFirstName());
-			array_push($authorText, $author->getMiddleName());
-			array_push($authorText, $author->getLastName());
-			$affiliations = $author->getAffiliation(null);
-			if (is_array($affiliations)) foreach ($affiliations as $affiliation) { // Localized
-				array_push($authorText, $affiliation);
+		// Check whether a search plug-in jumps in.
+		$hookResult =& HookRegistry::call(
+			'ArticleSearchIndex::indexArticleMetadata',
+			array($article)
+		);
+
+		// If no search plug-in is activated then fall back to the
+		// default database search implementation.
+		if ($hookResult === false || is_null($hookResult)) {
+			// Build author keywords
+			$authorText = array();
+			$authors = $article->getAuthors();
+			for ($i=0, $count=count($authors); $i < $count; $i++) {
+				$author =& $authors[$i];
+				array_push($authorText, $author->getFirstName());
+				array_push($authorText, $author->getMiddleName());
+				array_push($authorText, $author->getLastName());
+				$affiliations = $author->getAffiliation(null);
+				if (is_array($affiliations)) foreach ($affiliations as $affiliation) { // Localized
+					array_push($authorText, $affiliation);
+				}
+				$bios = $author->getBiography(null);
+				if (is_array($bios)) foreach ($bios as $bio) { // Localized
+					array_push($authorText, strip_tags($bio));
+				}
 			}
-			$bios = $author->getBiography(null);
-			if (is_array($bios)) foreach ($bios as $bio) { // Localized
-				array_push($authorText, strip_tags($bio));
-			}
+
+			// Update search index
+			$articleId = $article->getId();
+			$this->_updateTextIndex($articleId, ARTICLE_SEARCH_AUTHOR, $authorText);
+			$this->_updateTextIndex($articleId, ARTICLE_SEARCH_TITLE, $article->getTitle(null));
+			$this->_updateTextIndex($articleId, ARTICLE_SEARCH_ABSTRACT, $article->getAbstract(null));
+
+			$this->_updateTextIndex($articleId, ARTICLE_SEARCH_DISCIPLINE, (array) $article->getDiscipline(null));
+			$this->_updateTextIndex($articleId, ARTICLE_SEARCH_SUBJECT, array_merge(array_values((array) $article->getSubjectClass(null)), array_values((array) $article->getSubject(null))));
+			$this->_updateTextIndex($articleId, ARTICLE_SEARCH_TYPE, $article->getType(null));
+			$this->_updateTextIndex($articleId, ARTICLE_SEARCH_COVERAGE, array_merge(array_values((array) $article->getCoverageGeo(null)), array_values((array) $article->getCoverageChron(null)), array_values((array) $article->getCoverageSample(null))));
+			// FIXME Index sponsors too?
 		}
-
-		// Update search index
-		$articleId = $article->getId();
-		ArticleSearchIndex::_updateTextIndex($articleId, ARTICLE_SEARCH_AUTHOR, $authorText);
-		ArticleSearchIndex::_updateTextIndex($articleId, ARTICLE_SEARCH_TITLE, $article->getTitle(null));
-		ArticleSearchIndex::_updateTextIndex($articleId, ARTICLE_SEARCH_ABSTRACT, $article->getAbstract(null));
-
-		ArticleSearchIndex::_updateTextIndex($articleId, ARTICLE_SEARCH_DISCIPLINE, (array) $article->getDiscipline(null));
-		ArticleSearchIndex::_updateTextIndex($articleId, ARTICLE_SEARCH_SUBJECT, array_merge(array_values((array) $article->getSubjectClass(null)), array_values((array) $article->getSubject(null))));
-		ArticleSearchIndex::_updateTextIndex($articleId, ARTICLE_SEARCH_TYPE, $article->getType(null));
-		ArticleSearchIndex::_updateTextIndex($articleId, ARTICLE_SEARCH_COVERAGE, array_merge(array_values((array) $article->getCoverageGeo(null)), array_values((array) $article->getCoverageChron(null)), array_values((array) $article->getCoverageSample(null))));
-		// FIXME Index sponsors too?
 	}
 
 	/**
@@ -187,21 +217,31 @@ class ArticleSearchIndex {
 	 * @param $suppFile object
 	 */
 	function indexSuppFileMetadata(&$suppFile) {
-		// Update search index
-		$articleId = $suppFile->getArticleId();
-		ArticleSearchIndex::_updateTextIndex(
-			$articleId,
-			ARTICLE_SEARCH_SUPPLEMENTARY_FILE,
-			array_merge(
-				array_values((array) $suppFile->getTitle(null)),
-				array_values((array) $suppFile->getCreator(null)),
-				array_values((array) $suppFile->getSubject(null)),
-				array_values((array) $suppFile->getTypeOther(null)),
-				array_values((array) $suppFile->getDescription(null)),
-				array_values((array) $suppFile->getSource(null))
-			),
-			$suppFile->getFileId()
+		// Check whether a search plug-in jumps in.
+		$hookResult =& HookRegistry::call(
+			'ArticleSearchIndex::indexSuppFileMetadata',
+			array($suppFile)
 		);
+
+		// If no search plug-in is activated then fall back to the
+		// default database search implementation.
+		if ($hookResult === false || is_null($hookResult)) {
+			// Update search index
+			$articleId = $suppFile->getArticleId();
+			$this->_updateTextIndex(
+				$articleId,
+				ARTICLE_SEARCH_SUPPLEMENTARY_FILE,
+				array_merge(
+					array_values((array) $suppFile->getTitle(null)),
+					array_values((array) $suppFile->getCreator(null)),
+					array_values((array) $suppFile->getSubject(null)),
+					array_values((array) $suppFile->getTypeOther(null)),
+					array_values((array) $suppFile->getDescription(null)),
+					array_values((array) $suppFile->getSource(null))
+				),
+				$suppFile->getFileId()
+			);
+		}
 	}
 
 	/**
@@ -209,23 +249,33 @@ class ArticleSearchIndex {
 	 * @param $article Article
 	 */
 	function indexArticleFiles(&$article) {
-		// Index supplementary files
-		$fileDao =& DAORegistry::getDAO('SuppFileDAO');
-		$files =& $fileDao->getSuppFilesByArticle($article->getId());
-		foreach ($files as $file) {
-			if ($file->getFileId()) {
-				ArticleSearchIndex::updateFileIndex($article->getId(), ARTICLE_SEARCH_SUPPLEMENTARY_FILE, $file->getFileId());
-			}
-			ArticleSearchIndex::indexSuppFileMetadata($file);
-		}
-		unset($files);
+		// Check whether a search plug-in jumps in.
+		$hookResult =& HookRegistry::call(
+			'ArticleSearchIndex::indexArticleFiles',
+			array($article)
+		);
 
-		// Index galley files
-		$fileDao =& DAORegistry::getDAO('ArticleGalleyDAO');
-		$files =& $fileDao->getGalleysByArticle($article->getId());
-		foreach ($files as $file) {
-			if ($file->getFileId()) {
-				ArticleSearchIndex::updateFileIndex($article->getId(), ARTICLE_SEARCH_GALLEY_FILE, $file->getFileId());
+		// If no search plug-in is activated then fall back to the
+		// default database search implementation.
+		if ($hookResult === false || is_null($hookResult)) {
+			// Index supplementary files
+			$fileDao =& DAORegistry::getDAO('SuppFileDAO');
+			$files =& $fileDao->getSuppFilesByArticle($article->getId());
+			foreach ($files as $file) {
+				if ($file->getFileId()) {
+					$this->updateFileIndex($article->getId(), ARTICLE_SEARCH_SUPPLEMENTARY_FILE, $file->getFileId());
+				}
+				$this->indexSuppFileMetadata($file);
+			}
+			unset($files);
+
+			// Index galley files
+			$fileDao =& DAORegistry::getDAO('ArticleGalleyDAO');
+			$files =& $fileDao->getGalleysByArticle($article->getId());
+			foreach ($files as $file) {
+				if ($file->getFileId()) {
+					$this->updateFileIndex($article->getId(), ARTICLE_SEARCH_GALLEY_FILE, $file->getFileId());
+				}
 			}
 		}
 	}
@@ -235,14 +285,14 @@ class ArticleSearchIndex {
 	 */
 	function rebuildIndex($log = false) {
 		// Check whether a search plug-in jumps in.
-		$result =& HookRegistry::call(
+		$hookResult =& HookRegistry::call(
 			'ArticleSearchIndex::rebuildIndex',
 			array($log)
 		);
 
 		// If no search plug-in is activated then fall back to the
 		// default database search implementation.
-		if ($result === false || is_null($result)) {
+		if ($hookResult === false || is_null($hookResult)) {
 			// Clear index
 			if ($log) echo 'Clearing index ... ';
 			$searchDao =& DAORegistry::getDAO('ArticleSearchDAO');
@@ -264,8 +314,8 @@ class ArticleSearchIndex {
 				while (!$articles->eof()) {
 					$article =& $articles->next();
 					if ($article->getDateSubmitted()) {
-						ArticleSearchIndex::indexArticleMetadata($article);
-						ArticleSearchIndex::indexArticleFiles($article);
+						$this->indexArticleMetadata($article);
+						$this->indexArticleFiles($article);
 						$numIndexed++;
 					}
 					unset($article);
